move types and utils from components/location to utils
