+++ src/utils/location/LocationShareErrors.ts
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { _t } from "../../../languageHandler";
+import { _t } from "../../languageHandler";
 
 export enum LocationShareError {
     MapStyleUrlNotConfigured = 'MapStyleUrlNotConfigured',
