+++ src/utils/location/index.ts
@@ -14,6 +14,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+export * from './findMapStyleUrl';
 export * from './locationEventGeoUri';
+export * from './LocationShareErrors';
 export * from './map';
 export * from './parseGeoUri';
